// WEBUNBUILD/interface/src/features/Identity/types.ts
 export interface IIdentityData {
      type: string;
      dl: string;
      class: string;
      expiration_date: string;
      end_date: string;
      lastName: string;
      firstName: string;
      date_of_birth: string;
      rstr: string;
      sex: string;
      height: string;
      hair: string;
      weight: string;
      eye_color: string;
      address: string;
      photo: string;
 }